feat: Convert all assets to be associated to the configured Ethereum chain
