+++ core/types/snapshot_nodes.go
@@ -93,18 +93,6 @@ type PayloadProofOfWork struct {
 	LastPruningBlock uint64
 }
 
-type PayloadActiveAssets struct {
-	ActiveAssets *ActiveAssets
-}
-
-type PayloadPendingAssets struct {
-	PendingAssets *PendingAssets
-}
-
-type PayloadPendingAssetUpdates struct {
-	PendingAssetUpdates *PendingAssetUpdates
-}
-
 type PayloadBankingBridgeState struct {
 	BankingBridgeState *BankingBridgeState
 }
@@ -598,18 +586,6 @@ type EquityShareLP struct {
 	Share  num.Decimal
 	Avg    num.Decimal
 	VStake num.Decimal
-}
-
-type ActiveAssets struct {
-	Assets []*Asset
-}
-
-type PendingAssets struct {
-	Assets []*Asset
-}
-
-type PendingAssetUpdates struct {
-	Assets []*Asset
 }
 
 type BankingBridgeState struct {
@@ -1511,108 +1487,6 @@ func (*PayloadSpotLiquidityTarget) Namespace() SnapshotNamespace {
 
 func (p *PayloadSpotLiquidityTarget) Key() string {
 	return fmt.Sprintf("target:%v", p.Target.MarketId)
-}
-
-func PayloadActiveAssetsFromProto(paa *snapshot.Payload_ActiveAssets) *PayloadActiveAssets {
-	return &PayloadActiveAssets{
-		ActiveAssets: ActiveAssetsFromProto(paa.ActiveAssets),
-	}
-}
-
-func (p PayloadActiveAssets) IntoProto() *snapshot.Payload_ActiveAssets {
-	return &snapshot.Payload_ActiveAssets{
-		ActiveAssets: p.ActiveAssets.IntoProto(),
-	}
-}
-
-func (*PayloadActiveAssets) isPayload() {}
-
-func (p *PayloadActiveAssets) plToProto() interface{} {
-	return p.IntoProto()
-}
-
-func (*PayloadActiveAssets) Namespace() SnapshotNamespace {
-	return AssetsSnapshot
-}
-
-func (*PayloadActiveAssets) Key() string {
-	return "active"
-}
-
-func PayloadPendingAssetsFromProto(ppa *snapshot.Payload_PendingAssets) *PayloadPendingAssets {
-	return &PayloadPendingAssets{
-		PendingAssets: PendingAssetsFromProto(ppa.PendingAssets),
-	}
-}
-
-func (p PayloadPendingAssets) IntoProto() *snapshot.Payload_PendingAssets {
-	return &snapshot.Payload_PendingAssets{
-		PendingAssets: p.PendingAssets.IntoProto(),
-	}
-}
-
-func (*PayloadPendingAssets) isPayload() {}
-
-func (p *PayloadPendingAssets) plToProto() interface{} {
-	return p.IntoProto()
-}
-
-func (*PayloadPendingAssets) Key() string {
-	return "pending"
-}
-
-func (*PayloadPendingAssets) Namespace() SnapshotNamespace {
-	return AssetsSnapshot
-}
-
-func PayloadPendingAssetUpdatesFromProto(ppa *snapshot.Payload_PendingAssetUpdates) *PayloadPendingAssetUpdates {
-	return &PayloadPendingAssetUpdates{
-		PendingAssetUpdates: PendingAssetUpdatesFromProto(ppa.PendingAssetUpdates),
-	}
-}
-
-func (p PayloadPendingAssetUpdates) IntoProto() *snapshot.Payload_PendingAssetUpdates {
-	return &snapshot.Payload_PendingAssetUpdates{
-		PendingAssetUpdates: p.PendingAssetUpdates.IntoProto(),
-	}
-}
-
-func (*PayloadPendingAssetUpdates) isPayload() {}
-
-func (p *PayloadPendingAssetUpdates) plToProto() interface{} {
-	return p.IntoProto()
-}
-
-func (*PayloadPendingAssetUpdates) Key() string {
-	return "pending_updates"
-}
-
-func (*PayloadPendingAssetUpdates) Namespace() SnapshotNamespace {
-	return AssetsSnapshot
-}
-
-func (a PendingAssetUpdates) IntoProto() *snapshot.PendingAssetUpdates {
-	ret := &snapshot.PendingAssetUpdates{
-		Assets: make([]*vega.Asset, 0, len(a.Assets)),
-	}
-	for _, a := range a.Assets {
-		ret.Assets = append(ret.Assets, a.IntoProto())
-	}
-	return ret
-}
-
-func PendingAssetUpdatesFromProto(aa *snapshot.PendingAssetUpdates) *PendingAssetUpdates {
-	ret := PendingAssetUpdates{
-		Assets: make([]*Asset, 0, len(aa.Assets)),
-	}
-	for _, a := range aa.Assets {
-		pa, err := AssetFromProto(a)
-		if err != nil {
-			panic(err)
-		}
-		ret.Assets = append(ret.Assets, pa)
-	}
-	return &ret
 }
 
 func PayloadBankingBridgeStateFromProto(pbbs *snapshot.Payload_BankingBridgeState) *PayloadBankingBridgeState {
@@ -2590,54 +2464,6 @@ func (*PayloadStakingAccounts) Key() string {
 
 func (*PayloadStakingAccounts) Namespace() SnapshotNamespace {
 	return StakingSnapshot
-}
-
-func ActiveAssetsFromProto(aa *snapshot.ActiveAssets) *ActiveAssets {
-	ret := ActiveAssets{
-		Assets: make([]*Asset, 0, len(aa.Assets)),
-	}
-	for _, a := range aa.Assets {
-		aa, err := AssetFromProto(a)
-		if err != nil {
-			panic(err)
-		}
-		ret.Assets = append(ret.Assets, aa)
-	}
-	return &ret
-}
-
-func (a ActiveAssets) IntoProto() *snapshot.ActiveAssets {
-	ret := &snapshot.ActiveAssets{
-		Assets: make([]*vega.Asset, 0, len(a.Assets)),
-	}
-	for _, a := range a.Assets {
-		ret.Assets = append(ret.Assets, a.IntoProto())
-	}
-	return ret
-}
-
-func PendingAssetsFromProto(aa *snapshot.PendingAssets) *PendingAssets {
-	ret := PendingAssets{
-		Assets: make([]*Asset, 0, len(aa.Assets)),
-	}
-	for _, a := range aa.Assets {
-		pa, err := AssetFromProto(a)
-		if err != nil {
-			panic(err)
-		}
-		ret.Assets = append(ret.Assets, pa)
-	}
-	return &ret
-}
-
-func (a PendingAssets) IntoProto() *snapshot.PendingAssets {
-	ret := &snapshot.PendingAssets{
-		Assets: make([]*vega.Asset, 0, len(a.Assets)),
-	}
-	for _, a := range a.Assets {
-		ret.Assets = append(ret.Assets, a.IntoProto())
-	}
-	return ret
 }
 
 func BankingWithdrawalsFromProto(bw *snapshot.BankingWithdrawals) *BankingWithdrawals {
